+++ src/model/IHashTable.java
@@ -4,10 +4,10 @@ public interface IHashTable<T> {
 
     public void chainedHashInsert(PatientNode node);
 
-    public PatientNode chainedHashSearch(int key);
+    public PatientNode chainedHashSearch(long key);
 
-    public PatientNode chainedHashDelete(int key);
+    public PatientNode chainedHashDelete(long key);
 
-    public int hashFunction(int key);
+    public long hashFunction(long key);
 
 }
